% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{create_scenario_grid}
\alias{create_scenario_grid}
\title{Create Default Scenario Grid}
\usage{
create_scenario_grid(
  n_govs = 10,
  n_months = 12,
  noise_levels = c(0.5, 1),
  regimes = c("MCAR", "MAR", "MNAR"),
  pi_bases = c(0.5, 0.7, 0.9),
  mnar_omegas = c(0, -0.3, -0.6)
)
}
\arguments{
\item{n_govs}{Vector of governorate counts to test}

\item{n_months}{Vector of month counts to test}

\item{noise_levels}{Vector of noise levels to test}

\item{regimes}{Missingness regimes to test}

\item{pi_bases}{Baseline observation probabilities to test}

\item{mnar_omegas}{MNAR sensitivity parameters to test}
}
\value{
Data frame of scenarios with scenario_id
}
\description{
Helper function to create a standard scenario grid for simulation studies.
}
