% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag.R
\name{dummy_checchi_2017_parameters}
\alias{dummy_checchi_2017_parameters}
\title{Generate Dummy Parameters for Checchi et al. (2017) DAG}
\usage{
dummy_checchi_2017_parameters()
}
\value{
A data frame with columns:
\describe{
\item{from}{Name of the predictor node}
\item{to}{Name of the outcome node}
\item{effect_size}{Change in CMR (deaths per 10,000/day) per 1-unit increase in the predictor}
\item{lag}{Lag in days before the effect manifests}
}
}
\description{
Returns a data frame of plausible effect sizes and lags for use with the Checchi et al. (2017) mortality DAG.
The parameters represent estimated contributions to crude mortality rate (CMR) in deaths per 10,000 people per day.
}
\examples{
params <- dummy_checchi_2017_parameters()
head(params)

}
