% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_with_imputation}
\alias{fit_with_imputation}
\title{Fit Model with Multiple Imputation}
\usage{
fit_with_imputation(
  data,
  final_formula,
  vars_needed,
  n_imputations = 5,
  priors = NULL,
  chains = 4,
  cores = 1,
  iter = 2000,
  warmup = floor(iter/2),
  seed = NULL,
  control = list(),
  family = gaussian(),
  ...
)
}
\arguments{
\item{data}{Data frame with missing values}

\item{final_formula}{Model formula}

\item{vars_needed}{Variables needed for model}

\item{n_imputations}{Number of imputations}

\item{priors}{brms priors}

\item{chains}{Number of MCMC chains}

\item{cores}{Number of CPU cores}

\item{iter}{MCMC iterations}

\item{warmup}{Warmup iterations}

\item{seed}{Random seed}

\item{control}{Stan control parameters}

\item{family}{Model family}

\item{...}{Additional arguments for brm}
}
\value{
List of brmsfit objects (one per imputation)
}
\description{
Uses mice package for multiple imputation, fits model to each imputed dataset,
and pools results.
}
\keyword{internal}
