% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nurah.R
\docType{package}
\name{nurah}
\alias{nurah-package}
\alias{nurah}
\title{nurah: Testing causal models of mortality in crisis-affected populations}
\description{
nurah is an R package designed as a research compendium, focused on evaluating
the feasibility of estimating mortality in humanitarian crises. It utilises
statistical models, directed acyclic graphs (DAGs), and explores data
availability challenges to produce robust and actionable mortality estimates.
}
\author{
\strong{Maintainer}: OJ Watson \email{oj.watson@hotmail.co.uk} (\href{https://orcid.org/0000-0003-2374-0741}{ORCID})

}
