% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_mortality_negbin}
\alias{simulate_mortality_negbin}
\title{Simulate Mortality Using Negative Binomial Distribution}
\usage{
simulate_mortality_negbin(
  daily_data,
  dag,
  mortality_node,
  phi = 1,
  intercept = -8,
  noise_level = 1,
  n_days,
  n_regions
)
}
\arguments{
\item{daily_data}{Daily simulation data with population column}

\item{dag}{DAG object containing parameters}

\item{mortality_node}{Name of mortality outcome node}

\item{phi}{Overdispersion parameter (higher = less overdispersion; phi -> Inf gives Poisson)}

\item{intercept}{Baseline log-rate intercept}

\item{noise_level}{Noise scaling factor}

\item{n_days}{Number of days in simulation}

\item{n_regions}{Number of regions}
}
\value{
daily_data with mortality column added
}
\description{
Simulates mortality counts Y_it ~ NegBin(mu_it, phi) with:
log(mu_it) = log(N_it) + eta_it
eta_it = intercept + sum_j beta_j * parent_j(it) + random effects (optional)
}
\keyword{internal}
