% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag.R
\name{visualise_dag}
\alias{visualise_dag}
\title{Visualise a Directed Acyclic Graph (DAG) with Rounded‐Corner Boxes}
\usage{
visualise_dag(
  dag,
  use_dag_layout_attr = TRUE,
  node_color = "skyblue",
  label_size = 5
)
}
\arguments{
\item{dag}{A nurah_dag from \code{define_dag()} or \code{checchi_2017_dag()}.}

\item{use_dag_layout_attr}{Logical; if TRUE, use attr(dag, "layout") if present.}

\item{node_color}{Colour for node fill. Default \"skyblue\".}

\item{label_size}{Numeric cex for labels. Default 5}
}
\value{
Invisibly, the matrix of (x,y) coordinates used.
}
\description{
Plots a dag, using a predefined layout if available.
}
